refactor(settings): extract Settings.tsx subcomponents into components
