+++ src/pages/Settings/components/WebauthnCredentialItem.tsx
@@ -1,0 +1,234 @@
+import React, { FormEvent, KeyboardEvent, useCallback, useContext, useState } from 'react';
+import { Trans, useTranslation } from 'react-i18next';
+
+import StatusContext from '@/context/StatusContext';
+
+import { WebauthnCredential } from '../../../api/types';
+import { formatDate } from 'wallet-common';
+import type { WebauthnPrfEncryptionKeyInfo } from '../../../services/keystore';
+import { isPrfKeyV2 } from '../../../services/keystore';
+
+import DeletePopup from '../../../components/Popups/DeletePopup';
+import Button from '../../../components/Buttons/Button';
+import { Edit, RefreshCcw, Trash2 } from 'lucide-react';
+
+export function useWebauthnCredentialNickname(credential: WebauthnCredential): string {
+	const { t } = useTranslation();
+	if (credential) {
+		return credential.nickname || `${t('pageSettings.passkeyItem.unnamed')} ${credential.id.substring(0, 8)}`;
+	} else {
+		return "";
+	}
+}
+
+const WebauthnCredentialItem = ({
+	credential,
+	prfKeyInfo,
+	onDelete,
+	onRename,
+	onUpgradePrfKey,
+}: {
+	credential: WebauthnCredential,
+	prfKeyInfo: WebauthnPrfEncryptionKeyInfo,
+	onDelete?: false | (() => Promise<void>),
+	onRename: (credential: WebauthnCredential, nickname: string | null) => Promise<boolean>,
+	onUpgradePrfKey: (prfKeyInfo: WebauthnPrfEncryptionKeyInfo) => void,
+}) => {
+	const { isOnline } = useContext(StatusContext);
+	const [nickname, setNickname] = useState(credential.nickname || '');
+	const [editing, setEditing] = useState(false);
+	const { t } = useTranslation();
+	const currentLabel = useWebauthnCredentialNickname(credential);
+	const [submitting, setSubmitting] = useState(false);
+	const [isDeleteConfirmationOpen, setIsDeleteConfirmationOpen] = useState(false);
+	const [loading, setLoading] = useState(false);
+
+	const openDeleteConfirmation = () => setIsDeleteConfirmationOpen(true);
+	const closeDeleteConfirmation = () => setIsDeleteConfirmationOpen(false);
+
+	const handleDelete = async () => {
+		if (onDelete) {
+			setLoading(true);
+			await onDelete(); // Wait for the delete function to complete
+			setLoading(false);
+			closeDeleteConfirmation();
+		}
+	};
+
+	const onCancelEditing = useCallback(
+		() => {
+			setNickname(credential.nickname || '');
+			setEditing(false);
+		},
+		[credential.nickname],
+	);
+
+	const onKeyUp = useCallback(
+		(event: KeyboardEvent<HTMLInputElement>) => {
+			if (event.key === "Escape") {
+				onCancelEditing();
+			}
+		},
+		[onCancelEditing],
+	);
+
+	const onSubmit = async (event: FormEvent<HTMLFormElement>) => {
+		event.preventDefault();
+		setSubmitting(true);
+		try {
+			const result = await onRename(credential, nickname);
+			if (result) {
+				setEditing(false);
+			}
+		} finally {
+			setSubmitting(false);
+		}
+	};
+
+	const needsPrfUpgrade = prfKeyInfo && !isPrfKeyV2(prfKeyInfo);
+
+	return (
+		<form
+			className="mb-2 pl-4 px-4 py-2 border border-lm-gray-400 dark:border-dm-gray-600 rounded-lg flex flex-row flex-wrap gap-y-2 overflow-x-auto"
+			onSubmit={onSubmit}
+		>
+			<div className="grow">
+				{editing
+					? (
+						<>
+							<div className="flex items-center gap-2">
+								<p className="font-semibold dark:text-white">
+									{t('pageSettings.passkeyItem.nickname')}:&nbsp;
+								</p>
+								<input
+									className="w-36 px-3 py-2 bg-lm-gray-200 dark:bg-dm-gray-800 border border-lm-gray-600 dark:border-dm-gray-400 dark:text-white rounded-lg inputDarkModeOverride"
+
+									type="text"
+									placeholder={t('pageSettings.passkeyItem.nicknameInput')}
+									value={nickname}
+									onChange={(event) => setNickname(event.target.value)}
+									aria-label={t('pageSettings.passkeyItem.nicknameInputAriaLabel', { passkeyLabel: currentLabel })}
+									onKeyUp={onKeyUp}
+									disabled={submitting}
+								/>
+							</div>
+						</>
+					)
+					: (
+						<div className="flex items-center">
+							<p>
+								<span className="font-semibold dark:text-white">
+									{t('pageSettings.passkeyItem.nickname')}:&nbsp;
+								</span>
+								<span className="italic">
+									{currentLabel}
+								</span>
+							</p>
+						</div>
+					)
+				}
+				<p className='dark:text-white'>
+					<span className="font-semibold">
+						{t('pageSettings.passkeyItem.created')}:&nbsp;
+					</span>
+					{formatDate(credential.createTime)}
+				</p>
+				<p className='dark:text-white'>
+					<span className="font-semibold">
+						{t('pageSettings.passkeyItem.lastUsed')}:&nbsp;
+					</span>
+					{formatDate(credential.lastUseTime)}</p>
+				<p className='dark:text-white flex gap-3 items-center'>
+					<span className="font-semibold">
+						{t('pageSettings.passkeyItem.canEncrypt')}:&nbsp;
+					</span>
+					{credential.prfCapable ? t('pageSettings.passkeyItem.canEncryptYes') : t('pageSettings.passkeyItem.canEncryptNo')}
+					{needsPrfUpgrade
+						&& <span className="py-1 px-2 rounded bg-lm-orange dark:bg-dm-orange text-lm-gray-900 font-bold">{t('pageSettings.passkeyItem.needsPrfUpgrade')}</span>
+					}
+				</p>
+			</div>
+
+			<div className="items-start	flex gap-2">
+				{needsPrfUpgrade
+					&&
+					<Button
+						id="upgrade-prf-settings"
+						variant="outline"
+						type="button"
+						onClick={() => onUpgradePrfKey(prfKeyInfo)}
+						aria-label={t('pageSettings.passkeyItem.prfUpgradeAriaLabel', { passkeyLabel: currentLabel })}
+					>
+						<RefreshCcw size={18} /> {t('pageSettings.passkeyItem.prfUpgrade')}
+					</Button>
+				}
+
+				{editing
+					? (
+
+						<div className='flex gap-2'>
+							<Button
+								id="cancel-editing-settings"
+								onClick={onCancelEditing}
+								disabled={submitting}
+								ariaLabel={t('pageSettings.passkeyItem.cancelChangesAriaLabel', { passkeyLabel: currentLabel })}
+							>
+								{t('common.cancel')}
+							</Button>
+							<Button
+								id="save-editing-settings"
+								type="submit"
+								disabled={submitting}
+								variant="primary"
+							>
+								{t('common.save')}
+							</Button>
+						</div>
+					)
+					: (
+						<Button
+							id="rename-passkey"
+							onClick={() => setEditing(true)}
+							variant="primary"
+							disabled={!isOnline}
+							aria-label={t('pageSettings.passkeyItem.renameAriaLabel', { passkeyLabel: currentLabel })}
+							title={!isOnline ? t("common.offlineTitle") : ""}
+						>
+							<Edit size={18} className="mr-2" />
+							{t('pageSettings.passkeyItem.rename')}
+						</Button>
+					)
+				}
+
+				{onDelete && (
+					<Button
+						id="delete-passkey"
+						onClick={openDeleteConfirmation}
+						variant="delete"
+						disabled={!isOnline}
+						aria-label={t('pageSettings.passkeyItem.deleteAriaLabel', { passkeyLabel: currentLabel })}
+						title={!isOnline ? t("common.offlineTitle") : t("pageSettings.passkeyItem.deleteButtonTitleUnlocked", { passkeyLabel: currentLabel })}
+						additionalClassName='ml-2 py-3'
+					>
+						<Trash2 size={18} />
+					</Button>
+				)}
+				<DeletePopup
+					isOpen={isDeleteConfirmationOpen}
+					onConfirm={handleDelete}
+					onClose={closeDeleteConfirmation}
+					message={
+						<Trans
+							i18nKey="pageSettings.passkeyItem.messageDeletePasskey"
+							values={{ nickname: nickname }}
+							components={{ strong: <strong /> }}
+						/>
+					}
+					loading={loading}
+				/>
+			</div>
+		</form>
+	);
+};
+
+export default WebauthnCredentialItem;
